requirements.txt
asn1crypto==0.24.0
cached-property==1.4.3
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dnspython==1.15.0
exchangelib==1.10.1
future==0.16.0
idna==2.7
lxml==4.2.4
ntlm-auth==1.2.0
pycparser==2.18
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-ntlm==1.1.0
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
